some round issues in LOGINFO of constructors
git-svn-id: svn+ssh://svn.in2p3.fr/class@631 0e7d625b-0364-4367-a6be-d5be4a48d228

+++ source/trunk/src/Reactor.cxx
@@ -87,8 +87,8 @@ Reactor::Reactor(CLASSLogger* log,
 
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is fixed (for now)! "<< endl;
-	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetLifeTime()/cYear) << " year" << endl;
+	INFO << "\t Creation time set at \t " << ((double)GetCreationTime())/((double)cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetLifeTime())/((double)cYear) << " year" << endl;
 	INFO << "\t The Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW (with Full Power " << power << " and " << CapacityFactor << " capacity factor)"<< endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
 	
@@ -126,8 +126,8 @@ Reactor::Reactor(CLASSLogger* log,
 
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is not fixed (for now)! "<< endl;
-	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetLifeTime()/cYear) << " year" << endl;
+	INFO << "\t Creation time set at \t " << ((double)GetCreationTime())/((double)cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetLifeTime())/((double)cYear) << " year" << endl;
 	INFO << "\t The Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW (with Full Power " << Power << " and " << CapacityFactor << " capacity factor)"<< endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
 	
@@ -169,10 +169,10 @@ Reactor::Reactor(CLASSLogger* log, PhysicsModels* fueltypeDB, FabricationPlant* 
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is not fixed ! "<< endl;
 	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetLifeTime()/cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetLifeTime())/((double)cYear) << " year" << endl;
 	INFO << "\t The Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW (with Full Power " << Power << " and " << CapacityFactor << " capacity factor)"<< endl;
 	INFO << "\t Burn-Up at end of Cycle set at \t " << (double)(fBurnUp) << " GWj/t" << endl;
-	INFO << "\t The corresponding Cycle Time is\t " << (double)(fCycleTime/cYear) << " year" << endl;
+	INFO << "\t The corresponding Cycle Time is\t " << ((double)fCycleTime)/((double)cYear) << " year" << endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
 
 
@@ -210,9 +210,9 @@ Reactor::Reactor(CLASSLogger* log, PhysicsModels* 	fueltypeDB,
 
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is not fixed ! "<< endl;
-	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t The Cycle Time set at\t " << (double)(fCycleTime/cYear) << " year" << endl;
+	INFO << "\t Creation time set at \t " << ((double)GetCreationTime())/((double)cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetCreationTime())/((double)cYear) << " year" << endl;
+	INFO << "\t The Cycle Time set at\t " << ((double)fCycleTime)/((double)cYear) << " year" << endl;
 	INFO << "\t Burn-Up at end of Cycle set at \t " << (double)(fBurnUp) << " GWj/t" << endl;
 	INFO << "\t The corresponding Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW" << endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
@@ -266,8 +266,8 @@ Reactor::Reactor(CLASSLogger* log, EvolutionData* evolutivedb,
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is fixed ! "<< endl;
 	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetLifeTime()/cYear) << " year" << endl;
-	INFO << "\t The Cycle Time set at\t " << (double)(fCycleTime/cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetLifeTime())/((double)cYear) << " year" << endl;
+	INFO << "\t The Cycle Time set at\t " << ((double)fCycleTime)/((double)cYear) << " year" << endl;
 	INFO << "\t The Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW (with Full Power " << power << " and " << CapacityFactor << " capacity factor)"<< endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
 
@@ -315,9 +315,9 @@ Reactor::Reactor(CLASSLogger* log, EvolutionData* evolutivedb,
 
 	INFO << " A Reactor has been define :" << endl;
 	INFO << "\t Fuel Composition is fixed ! "<< endl;
-	INFO << "\t Creation time set at \t " << (double)(GetCreationTime()/cYear) << " year" << endl;
-	INFO << "\t Life time (Operating's Duration) set at \t " << (double)(GetLifeTime()/cYear) << " year" << endl;
-	INFO << "\t The Cycle Time set at\t " << (double)(fCycleTime/cYear) << " year" << endl;
+	INFO << "\t Creation time set at \t " << ((double)GetCreationTime())/((double)cYear) << " year" << endl;
+	INFO << "\t Life time (Operating's Duration) set at \t " << ((double)GetLifeTime())/((double)cYear) << " year" << endl;
+	INFO << "\t The Cycle Time set at\t " << ((double)fCycleTime)/((double)cYear) << " year" << endl;
 	INFO << "\t The Effective Thermal Power is \t " << (double)(fPower *1e-6) << " MW (with Full Power " << fPower << endl;
 	INFO << "\t The Heavy Metal Mass in the Core set at " << (double)(fHeavyMetalMass) << " tons" << endl << endl;
